Animation Script addition and Patrol updated
Untested but the animation script should in theory work for every npc regardless of what other scripts are attached. As in regardless if they move because of say Guard script or Patrol script, the animations should run properly. But like I said, nothing is tested and once the other scripts are made we will know for sure.